Move TaskNodeProvider to Providers folder
