+++ src/components/shared/ReactFlow/FlowCanvas/TaskNode/TaskNodeCard/Handles.tsx
@@ -7,9 +7,8 @@ import {
 import { useEffect, useRef } from "react";
 
 import { cn } from "@/lib/utils";
+import { useTaskNode } from "@/providers/TaskNodeProvider";
 import type { InputSpec, OutputSpec } from "@/utils/componentSpec";
-
-import { useTaskNode } from "../TaskNodeProvider";
 
 type InputHandleProps = {
   input: InputSpec;
